requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
